+++ .storybook/main.ts
@@ -15,7 +15,6 @@ const config: StorybookConfig = {
   stories: ['../packages/**/src/**/*.stories.@(js|jsx|ts|tsx)'],
   addons: [
     getAbsolutePath('storybook-addon-tag-badges'),
-    getAbsolutePath('@storybook/addon-onboarding'),
     getAbsolutePath('@storybook/addon-links'),
     getAbsolutePath('@storybook/addon-essentials'),
     getAbsolutePath('@storybook/addon-interactions'),