+++ packages/react-components/src/components/sheet/sheet.module.css.d.ts
@@ -1,0 +1,11 @@
+declare const styles: {
+  readonly "Container": string;
+  readonly "HandleWrapper": string;
+  readonly "Header": string;
+  readonly "Overlay": string;
+  readonly "SafeGuard": string;
+  readonly "Scroller": string;
+  readonly "Sheet": string;
+};
+export = styles;
+