+++ packages/react-components/src/components/sheet/sheet.tsx
@@ -1,0 +1,234 @@
+'use client';
+
+import {
+  FC, forwardRef, ReactNode, useMemo,
+} from 'react';
+import { DialogProps, Drawer as Vaul } from 'vaul';
+
+import {
+  ClampText, Panel, ResponsiveProvider, ScrollArea,
+  Stack, Text, Title,
+  useResponsiveContext,
+} from '@/components';
+
+import styles from './sheet.module.css';
+
+type SheetContentProps = {
+  /**
+   * The interactive element that triggers the modal to open.
+   * Must be a single interactive element.
+   */
+  trigger: React.ComponentProps<typeof Vaul.Trigger>['children'];
+  /**
+   * Set the accessible title of the modal. This is used by screen readers to
+   * announce the title of the modal when opened.
+   */
+  heading: ReactNode;
+  /**
+   * Hide or show the sheet heading which includes the title and the description.
+   * If you hide the title, make sure to add yuor own title element in order
+   * to provide an accessible label to the drawer.
+   *
+   * @defaultValue true
+   */
+  showHeading?: boolean;
+  /**
+   * Set the background color for the content header
+   */
+  headerTint?: string;
+  /**
+   * An optional accessible description to be announced when the drawer is opened.
+   */
+  description?: React.ComponentProps<typeof Vaul.Description>['children'];
+  /**
+   * Enable or disable safe padding. Prevents content from being hidden behind the safe area.
+   *
+   * @defaultValue true
+   */
+  safePadding?: boolean;
+  /**
+   * Enable or disable padding for the content.
+   */
+  noPadding?: boolean;
+  /**
+   * Set the header of the drawer to be compact to save vertical space.
+   * This is useful when the drawer is used in a mobile context.
+   *
+   * !important: Make sure to pass a title that fits in the compact header.
+   *
+   * @defaultValue false
+   */
+  compactHeader?: boolean;
+  /**
+   * Set the sheet's header to be sticky when scrolling.
+   */
+  stickyHeader?: boolean;
+  /**
+   * Set the maximum width of the drawer, on mobile the drawer will be
+   * automatically collapsed when the width is reached.
+   */
+  maxWidth?: string;
+  /**
+   * Ref to the drawer content
+   *
+   */
+  scrollerRef?: React.RefObject<HTMLDivElement>;
+}
+
+const SheetContent = forwardRef<
+HTMLDivElement,
+SheetContentProps & Pick<DialogProps, 'children' | 'dismissible' | 'direction'>
+>(({
+  trigger,
+  heading,
+  headerTint,
+  description,
+  safePadding = true,
+  noPadding = false,
+  children,
+  dismissible = true,
+  direction = 'bottom',
+  maxWidth = '600px',
+  compactHeader,
+  stickyHeader = false,
+  showHeading = true,
+  scrollerRef,
+}, forwardedRef) => {
+  const { matches } = useResponsiveContext();
+
+  const align = useMemo(() => {
+    if (direction === 'left') return 'start';
+    if (direction === 'right') return 'end';
+    return 'center';
+  }, [direction]);
+
+  const dynamicStyle = useMemo(() => (
+    {
+      '--max-w': maxWidth ?? undefined,
+      '--header-tint': headerTint,
+    }
+  ), [headerTint, maxWidth]);
+
+  const getVPadding = useMemo(() => {
+    if (showHeading) return 0;
+    return noPadding ? 0 : 16;
+  }, [showHeading, noPadding]);
+
+  return (
+    <>
+      <Vaul.Trigger asChild>
+        {trigger}
+      </Vaul.Trigger>
+      <Vaul.Portal>
+        <Vaul.Overlay className={styles.Overlay} />
+        <Vaul.Content asChild>
+          <Stack
+            style={{ ...dynamicStyle }}
+            className={styles.Sheet}
+            tabIndex={-1}
+            ref={forwardedRef}
+            hAlign={align}
+            vAlign="start"
+            hPadding={8}
+            vPadding={8}
+          >
+            <Panel
+              vibrant
+              vibrancyColor="soft"
+              bordered
+              showGlow
+              radius={24}
+              glowFitContent
+            >
+
+              <Stack className={styles.Container}>
+                <ScrollArea
+                  className={styles.Scroller}
+                  useSystemStyle={false}
+                  hideScrollbars={!matches.small}
+                  ref={scrollerRef}
+                >
+                  {/* Drag Handle */}
+                  {!matches.small && (
+                    <>
+                      {dismissible && (
+                        <Stack vPadding={8} className={styles.HandleWrapper}>
+                          <Vaul.Handle />
+                        </Stack>
+                      )}
+                    </>
+                  )}
+
+                  {/* Header */}
+                  <Stack
+                    rowGap={4}
+                    hPadding={24}
+                    vPadding={compactHeader ? 8 : 24}
+                    className={styles.Header}
+                    data-sheet-sticky-header={stickyHeader}
+                    data-visually-hidden={!showHeading}
+                  >
+                    <Vaul.Title asChild>
+                      <Title lineHeight="small" responsive={false} level={compactHeader ? '6' : '5'}>{heading}</Title>
+                    </Vaul.Title>
+
+                    {description && (
+                      <Vaul.Description asChild>
+                        <Text dimmed={5} weight="regular" size={compactHeader ? 14 : 16}>
+                          <ClampText rows={2}>{description}</ClampText>
+                        </Text>
+                      </Vaul.Description>
+                    )}
+                  </Stack>
+
+                  {/* Sheet content */}
+                  <Stack
+                    hPadding={noPadding ? undefined : 24}
+                    vPadding={[getVPadding, 0]}
+                    className={styles.SafeGuard}
+                    data-modal-content-safe-padding={noPadding ? false : safePadding}
+                  >
+                    {children}
+                  </Stack>
+                </ScrollArea>
+              </Stack>
+
+            </Panel>
+          </Stack>
+        </Vaul.Content>
+      </Vaul.Portal>
+    </>
+  );
+});
+
+export const SheetWrapper: FC<SheetProps> = ({
+  nested,
+  ...otherProps
+}) => {
+  const { matches } = useResponsiveContext();
+
+  return nested ? (
+    <Vaul.NestedRoot
+      {...otherProps}
+      handleOnly={matches.small}
+    >
+      <SheetContent {...otherProps} />
+    </Vaul.NestedRoot>
+  ) : (
+    <Vaul.Root
+      {...otherProps}
+      handleOnly={matches.small}
+    >
+      <SheetContent {...otherProps} />
+    </Vaul.Root>
+  );
+};
+
+export type SheetProps = DialogProps & SheetContentProps & { nested?: boolean };
+
+export const Sheet: FC<SheetProps> = ({ nested, ...otherProps }) => (
+  <ResponsiveProvider>
+    <SheetWrapper nested={nested} {...otherProps} />
+  </ResponsiveProvider>
+);
+