+++ packages/react-components/src/components/stack/stack.tsx
@@ -49,11 +49,11 @@ export type StackProps = {
   /**
    * Set the horizontal padding (left/right)
    */
-  hPadding?: TokensTypes['space'];
+  hPadding?: TokensTypes['space'] | [TokensTypes['space'] | 0, TokensTypes['space'] | 0];
   /**
    * Set the vertical padding (top/bottom)
    */
-  vPadding?: TokensTypes['space'];
+  vPadding?: TokensTypes['space'] | [TokensTypes['space'] | 0, TokensTypes['space'] | 0];
   /**
    * Renders children as rows or columns. The value can be one of the flex directions.
    * More info: https://developer.mozilla.org/en-US/docs/Web/CSS/flex-direction
@@ -92,16 +92,31 @@ export const Stack = forwardRef(
       return prop;
     };
 
-    const dynamicStyle = useMemo(() => (
-      {
+    const dynamicStyle = useMemo(() => {
+      const getPaddingValue = (
+        padding: TokensTypes['space'] | [TokensTypes['space'] | 0, TokensTypes['space'] | 0],
+        direction: 'horizontal' | 'vertical',
+      ) => {
+        const [start, end] = Array.isArray(padding) ? padding : [padding, padding];
+        return direction === 'horizontal'
+          ? { left: start ? tkns.space[start] : 0, right: end ? tkns.space[end] : 0 }
+          : { top: start ? tkns.space[start] : 0, bottom: end ? tkns.space[end] : 0 };
+      };
+
+      const vPaddingValues = vPadding ? getPaddingValue(vPadding, 'vertical') : { top: 0, bottom: 0 };
+      const hPaddingValues = hPadding ? getPaddingValue(hPadding, 'horizontal') : { left: 0, right: 0 };
+
+      return {
         '--r-gap': rowGap ? tkns.space[rowGap] : 0,
         '--c-gap': columnGap ? tkns.space[columnGap] : 0,
-        '--v-align': vAlign && alignmentTemplate(vAlign),
-        '--h-align': hAlign && alignmentTemplate(hAlign),
-        '--v-padding': vPadding ? tkns.space[vPadding] : 0,
-        '--h-padding': hPadding ? tkns.space[hPadding] : 0,
-      }
-    ), [columnGap, hAlign, hPadding, rowGap, vAlign, vPadding]);
+        '--v-align': alignmentTemplate(vAlign),
+        '--h-align': alignmentTemplate(hAlign),
+        '--v-padding-top': vPaddingValues.top,
+        '--v-padding-bottom': vPaddingValues.bottom,
+        '--h-padding-left': hPaddingValues.left,
+        '--h-padding-right': hPaddingValues.right,
+      };
+    }, [columnGap, hAlign, hPadding, rowGap, vAlign, vPadding]);
 
     return (
       <Component