+++ packages/charts/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @lualtek/charts
 
+## 18.5.0-next.6
+
+### Patch Changes
+
+- Updated dependencies [258c097]
+  - @lualtek/react-components@18.5.0-next.6
+  - @lualtek/tokens@18.5.0-next.6
+
+## 18.5.0-next.5
+
+### Patch Changes
+
+- Updated dependencies [bfac287]
+  - @lualtek/react-components@18.5.0-next.5
+  - @lualtek/tokens@18.5.0-next.5
+
+## 18.5.0-next.4
+
+### Patch Changes
+
+- Updated dependencies [ea30fe5]
+  - @lualtek/react-components@18.5.0-next.4
+  - @lualtek/tokens@18.5.0-next.4
+
+## 18.5.0-next.3
+
+### Patch Changes
+
+- 8b96840: Minor fixes
+  - @lualtek/react-components@18.5.0-next.3
+  - @lualtek/tokens@18.5.0-next.3
+
+## 18.5.0-next.2
+
+### Patch Changes
+
+- Updated dependencies [1555737]
+  - @lualtek/react-components@18.5.0-next.2
+  - @lualtek/tokens@18.5.0-next.2
+
+## 18.5.0-next.1
+
+### Patch Changes
+
+- Updated dependencies [4fb0e88]
+  - @lualtek/react-components@18.5.0-next.1
+  - @lualtek/tokens@18.5.0-next.1
+
+## 18.5.0-next.0
+
+### Patch Changes
+
+- Updated dependencies [842beb1]
+- Updated dependencies [e4b4452]
+  - @lualtek/react-components@18.5.0-next.0
+  - @lualtek/tokens@18.5.0-next.0
+
 ## 18.4.0
 
 ### Patch Changes